Created hierarchical groups for admin
Needed this to get groups able to "nest" for each division. In the
future will give us the ability to check if the parent exists and then
take the leaders from that group and add them as leaders to the child
group. This will allow for a trickle down of all the leaders through the
groups. Will ultimately be able to do this for authors as well.