Updated Learndash plugin
Tested it and seems like it works fine. Replaced the
class-learndash-admin-groups-edit.php with the old version to keep the
changes made for Fulltime & Quebec courses. Also kep the ld-groups.php
file the same as the old version. The key thing here was that it was
looking for the role 'group_leader' and I need to to look for the
capability 'group-leader'.