Terminated User Redirect
When you set a user as terminated it now will redirect the user to a
"disabled" page with a message saying "This account has currently been
disabled. If you believe this is a mistake please contact your
supervisor to re-activate your account.".

This is rather than deleting the user file in the system and therefore
maintains the information.